+++ consensus/sync_converge.go
@@ -9,15 +9,15 @@ import (
 	"github.com/ngchain/ngcore/utils"
 )
 
-// MustFork detection ignites the forking in local node
+// MustConverge detection ignites the forking in local node
 // then do a filter covering all remotes to get the longest chain (if length is same, choose the heavier latest block one)
-func (mod *syncModule) MustFork(slice []*RemoteRecord) []*RemoteRecord {
+func (mod *syncModule) MustConverge(slice []*RemoteRecord) []*RemoteRecord {
 	var ret = make([]*RemoteRecord, 0)
 	latestHeight := mod.pow.Chain.GetLatestBlockHeight()
 	latestCheckPoint := mod.pow.Chain.GetLatestCheckpoint()
 
 	for _, r := range slice {
-		if r.shouldFork(latestCheckPoint, latestHeight) {
+		if r.shouldConverge(latestCheckPoint, latestHeight) {
 			ret = append(ret, r)
 		}
 	}
@@ -26,8 +26,8 @@ func (mod *syncModule) MustFork(slice []*RemoteRecord) []*RemoteRecord {
 }
 
 // force local chain be same as the remote record
-// fork is a danger operation so all msg are warn level
-func (mod *syncModule) doFork(record *RemoteRecord) error {
+// converge is a danger operation so all msg are warn level
+func (mod *syncModule) doConverging(record *RemoteRecord) error {
 	if mod.Locker.IsLocked() {
 		return nil
 	}
@@ -35,14 +35,14 @@ func (mod *syncModule) doFork(record *RemoteRecord) error {
 	mod.Locker.Lock()
 	defer mod.Locker.Unlock()
 
-	log.Warnf("start forking Chain from remote node %s, target height: %d", record.id, record.latest)
-	chain, err := mod.getBlocksSinceForkPoint(record)
+	log.Warnf("start converging chain from remote node %s, target height: %d", record.id, record.latest)
+	chain, err := mod.getBlocksSinceDiffPoint(record)
 	if err != nil {
 		return err
 	}
 
 	localSamepoint, _ := mod.pow.Chain.GetBlockByHeight(chain[1].Height)
-	log.Warnf("have got the fork point: block@%d: local: %x remote %x", chain[1].Height, chain[1].Hash(), localSamepoint.Hash())
+	log.Warnf("have got the diffpoint: block@%d: local: %x remote %x", chain[1].Height, chain[1].Hash(), localSamepoint.Hash())
 
 	err = mod.pow.Chain.ForceApplyBlocks(chain)
 	if err != nil {
@@ -60,12 +60,12 @@ func (mod *syncModule) doFork(record *RemoteRecord) error {
 		return err
 	}
 
-	log.Warn("fork finished")
+	log.Warn("converging finished")
 	return nil
 }
 
-// getBlocksSinceForkPoint gets the fork point by comparing hashes between local and remote
-func (mod *syncModule) getBlocksSinceForkPoint(record *RemoteRecord) ([]*ngtypes.Block, error) {
+// getBlocksSinceDiffPoint gets the diffpoint by comparing hashes between local and remote
+func (mod *syncModule) getBlocksSinceDiffPoint(record *RemoteRecord) ([]*ngtypes.Block, error) {
 	blocks := make([]*ngtypes.Block, 0)
 	blockHashes := make([][]byte, defaults.MaxBlocks)
 
@@ -76,7 +76,7 @@ func (mod *syncModule) getBlocksSinceForkPoint(record *RemoteRecord) ([]*ngtypes
 	// when the chainLen (the len of returned chain) is not equal to defaults.MaxBlocks, means it has reach the latest height
 	for {
 		if ptr == 0 {
-			panic("cannot fork: completely different chains!")
+			panic("converging failed: completely different chains!")
 		}
 
 		to := ptr
@@ -95,7 +95,7 @@ func (mod *syncModule) getBlocksSinceForkPoint(record *RemoteRecord) ([]*ngtypes
 			blockHashes[h-ptr-1] = b.Hash()
 		}
 
-		// To == from+to means fork mode
+		// To == from+to means converging mode
 		chain, err := mod.getRemoteChain(record.id, blockHashes, bytes.Join([][]byte{utils.PackUint64LE(from), utils.PackUint64LE(to)}, nil))
 		if err != nil {
 			return nil, err